# openapi/v1/api.yaml
openapi: 3.0.0
paths:
...
  /v1/tenants/current:
    get:
      operationId: TenantsController_getCurrent
      summary: Obter dados da empresa atual
      parameters: []
      responses:
        '200':
          description: Dados da empresa
        '401':
          description: Não autenticado
        '403':
          description: Sem permissão
      tags:
        - tenants
      security:
        - bearer: []
    patch:
      operationId: TenantsController_update
      summary: Atualizar dados da empresa atual
      parameters: []
      responses:
        '200':
          description: Empresa atualizada com sucesso
        '400':
          description: Dados inválidos
        '401':
          description: Não autenticado
        '403':
          description: Sem permissão
      tags:
        - tenants
      security:
        - bearer: []
  /v1/tenants/branches:
    get:
      operationId: TenantsController_getBranches
      summary: Listar filiais da empresa
      parameters: []
      responses:
        '200':
          description: Lista de filiais
        '401':
          description: Não autenticado
        '403':
          description: Sem permissão
      tags:
        - tenants
      security:
        - bearer: []
    post:
      operationId: TenantsController_createBranch
      summary: Criar nova filial
      parameters: []
      responses:
        '201':
          description: Filial criada com sucesso
        '400':
          description: Dados inválidos
        '401':
          description: Não autenticado
        '403':
          description: Sem permissão
      tags:
        - tenants
      security:
        - bearer: []
  /v1/tenants/branches/{id}:
    patch:
      operationId: TenantsController_updateBranch
      summary: Atualizar filial
      parameters:
        - name: id
          required: true
          in: path
          schema:
            type: string
      responses:
        '200':
          description: Filial atualizada com sucesso
        '400':
          description: Dados inválidos
        '401':
          description: Não autenticado
        '403':
          description: Sem permissão
        '404':
          description: Filial não encontrada
      tags:
        - tenants
      security:
        - bearer: []
...
  /v1/roles:
    get:
      operationId: RolesController_findAll
      summary: Listar cargos da empresa
      parameters: []
      responses:
        '200':
          description: Lista de cargos
        '401':
          description: Não autenticado
        '403':
          description: Sem permissão
      tags:
        - roles
      security:
        - bearer: []
    post:
      operationId: RolesController_create
      summary: Criar novo cargo
      parameters: []
      responses:
        '201':
          description: Cargo criado com sucesso
        '400':
          description: Dados inválidos
        '401':
          description: Não autenticado
        '403':
          description: Sem permissão
      tags:
        - roles
      security:
        - bearer: []
  /v1/roles/{id}:
    get:
      operationId: RolesController_findById
      summary: Obter cargo por ID
      parameters:
        - name: id
          required: true
          in: path
          schema:
            type: string
      responses:
        '200':
          description: Dados do cargo com permissões
        '401':
          description: Não autenticado
        '403':
          description: Sem permissão
        '404':
          description: Cargo não encontrado
      tags:
        - roles
      security:
        - bearer: []
    patch:
      operationId: RolesController_update
      summary: Atualizar cargo
      parameters:
        - name: id
          required: true
          in: path
          schema:
            type: string
      responses:
        '200':
          description: Cargo atualizado com sucesso
        '400':
          description: Dados inválidos
        '401':
          description: Não autenticado
        '403':
          description: Sem permissão
        '404':
          description: Cargo não encontrado
      tags:
        - roles
      security:
        - bearer: []
    delete:
      operationId: RolesController_softDelete
      summary: Excluir cargo (soft delete)
      parameters:
        - name: id
          required: true
          in: path
          schema:
            type: string
      responses:
        '204':
          description: Cargo excluído com sucesso
        '401':
          description: Não autenticado
        '403':
          description: Sem permissão ou cargo do sistema
        '404':
          description: Cargo não encontrado
      tags:
        - roles
      security:
        - bearer: []
  /v1/roles/{id}/permissions:
    post:
      operationId: RolesController_assignPermission
      summary: Atribuir permissão ao cargo
      parameters:
        - name: id
          required: true
          in: path
          schema:
            type: string
      responses:
        '201':
          description: Permissão atribuída com sucesso
        '400':
          description: Dados inválidos
        '401':
          description: Não autenticado
        '403':
          description: Sem permissão
        '404':
          description: Cargo ou permissão não encontrada
      tags:
        - roles
      security:
        - bearer: []
  /v1/roles/{roleId}/permissions/{permissionId}:
    delete:
      operationId: RolesController_removePermission
      summary: Remover permissão do cargo
      parameters:
        - name: roleId
          required: true
          in: path
          schema:
            type: string
        - name: permissionId
          required: true
          in: path
          schema:
            type: string
      responses:
        '204':
          description: Permissão removida com sucesso
        '401':
          description: Não autenticado
        '403':
          description: Sem permissão
        '404':
          description: Cargo ou permissão não encontrada
      tags:
        - roles
      security:
        - bearer: []
...
  /v1/routes:
    get:
      operationId: RoutesController_findAll
      summary: List all routes with cursor-based pagination
      parameters: []
      responses:
        '200':
          description: ''
      tags:
        - routes
      security:
        - bearer: []
    post:
      operationId: RoutesController_create
      summary: Create a new route
      parameters: []
      responses:
        '201':
          description: ''
      tags:
        - routes
      security:
        - bearer: []
  /v1/routes/{id}:
    get:
      operationId: RoutesController_findById
      summary: Get route by ID with route points
      parameters:
        - name: id
          required: true
          in: path
          schema:
            type: string
      responses:
        '200':
          description: ''
      tags:
        - routes
      security:
        - bearer: []
    patch:
      operationId: RoutesController_update
      summary: Update route
      parameters:
        - name: id
          required: true
          in: path
          schema:
            type: string
      responses:
        '200':
          description: ''
      tags:
        - routes
      security:
        - bearer: []
    delete:
      operationId: RoutesController_delete
      summary: Soft delete a route
      parameters:
        - name: id
          required: true
          in: path
          schema:
            type: string
      responses:
        '204':
          description: ''
      tags:
        - routes
      security:
        - bearer: []
  /v1/routes/{id}/points:
    patch:
      operationId: RoutesController_updatePoints
      summary: Replace all route points
      parameters:
        - name: id
          required: true
          in: path
          schema:
            type: string
      responses:
        '200':
          description: ''
      tags:
        - routes
      security:
        - bearer: []
...
info:
  title: Gestão Fretamento Pro API
  description: API REST para gestão de fretamento empresarial
  version: 1.0.0
  contact: {}
...
servers: []
components:
  securitySchemes:
    bearer:
      scheme: bearer
      bearerFormat: JWT
      type: http
...